requirements.txt
Django==4.0
django-allauth==0.45.0
django-bootstrap4==3.0.1
django-crispy-forms==2.0
django-dotenv==1.4.2
django-environ==0.4.5
django-extensions==3.2.1
django-localflavor==3.1
django-stdimage==5.3.0
django-storages==1.11.1
gunicorn==20.1.0
python-dotenv==0.21.0
psycopg2-binary==2.9.5
dj-database-url==1.2.0
whitenoise==6.2.0
crispy-bootstrap4==2022.1
djangorestframework==3.14.0
importlib-metadata==6.8.0 
markdown==3.4.4 
zipp==3.16.2
django-filter==23.2